@@ -2,11 +2,13 @@
   "name": "complete-intro-to-react",
   "version": "1.0.0",
   "description": "A two day workshop on a complete intro to React",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "test": "mocha --require test/helpers/setup.js",
+    "start": "node app.js",
     "build": "webpack",
     "watch": "webpack --watch",
+    "test": "mocha --require test/helpers/setup.js",
+    "lint": "eslint --ignore-path .gitignore --cache ./",
     "cover": "nyc --reporter=lcov --reporter=text --reporter=html --require babel-register --extension .jsx npm test"
   },
   "author": "Brian Holt <btholt@gmail.com>",
@@ -15,36 +17,37 @@
     "axios": "0.9.1",
     "express": "4.13.4",
     "lodash": "4.6.1",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-redux": "^4.4.0",
-    "react-router": "^2.0.0",
-    "redux": "^3.3.1"
+    "node-ensure": "0.0.0",
+    "react": "0.14.7",
+    "react-dom": "0.14.7",
+    "react-redux": "4.4.0",
+    "react-router": "2.0.0",
+    "redux": "3.3.1"
   },
   "devDependencies": {
-    "babel-core": "^6.5.2",
-    "babel-loader": "^6.2.2",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-register": "^6.5.2",
-    "chai": "^3.5.0",
-    "enzyme": "^2.0.0",
-    "eslint": "^2.2.0",
-    "eslint-config-standard": "^5.1.0",
-    "eslint-config-standard-jsx": "^1.1.1",
-    "eslint-config-standard-react": "^2.3.0",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-promise": "^1.0.8",
-    "eslint-plugin-react": "^4.1.0",
-    "eslint-plugin-standard": "^1.3.2",
-    "jsdom": "^8.0.4",
-    "json-loader": "^0.5.4",
-    "mocha": "^2.4.5",
-    "nyc": "^6.0.0",
-    "react-addons-test-utils": "^0.14.7",
-    "sinon": "^1.17.3",
-    "webpack": "^1.12.13"
+    "babel-core": "6.5.2",
+    "babel-loader": "6.2.2",
+    "babel-polyfill": "6.7.4",
+    "babel-preset-es2015": "6.5.0",
+    "babel-preset-react": "6.5.0",
+    "babel-register": "6.5.2",
+    "chai": "3.5.0",
+    "enzyme": "2.0.0",
+    "eslint": "2.2.0",
+    "eslint-config-standard": "5.1.0",
+    "eslint-config-standard-jsx": "1.1.1",
+    "eslint-config-standard-react": "2.3.0",
+    "eslint-loader": "1.3.0",
+    "eslint-plugin-promise": "1.0.8",
+    "eslint-plugin-react": "4.1.0",
+    "eslint-plugin-standard": "1.3.2",
+    "jsdom": "8.0.4",
+    "json-loader": "0.5.4",
+    "mocha": "2.4.5",
+    "nyc": "6.0.0",
+    "react-addons-test-utils": "0.14.7",
+    "sinon": "1.17.3",
+    "webpack": "1.13.2"
   },
   "directories": {
     "test": "test"
