@@ -1,12 +1,12 @@
 const React = require('react')
 const ShowCard = require('./ShowCard')
 const Header = require('./Header')
-const { object, string } = React.PropTypes
+const { arrayOf, object, string } = React.PropTypes
 const { connector } = require('./Store')
 
 const Search = React.createClass({
   propTypes: {
-    route: object,
+    shows: arrayOf(object),
     searchTerm: string
   },
   render () {
@@ -14,7 +14,7 @@ const Search = React.createClass({
       <div className='container'>
         <Header showSearch />
         <div className='shows'>
-          {this.props.route.shows
+          {this.props.shows
             .filter((show) => `${show.title} ${show.description}`.toUpperCase().indexOf(this.props.searchTerm.toUpperCase()) >= 0)
             .map((show) => (
               <ShowCard {...show} key={show.imdbID} />
