@@ -1,31 +1,21 @@
 const React = require('react')
-const ReactDOM = require('react-dom')
 const Landing = require('./Landing')
 const Search = require('./Search')
 const Layout = require('./Layout')
 const Details = require('./Details')
 const { Router, Route, IndexRoute, hashHistory } = require('react-router')
-const { shows } = require('../public/data')
 const { store } = require('./Store')
 const { Provider } = require('react-redux')
 
 const App = React.createClass({
-  assignShow (nextState, replace) {
-    const showArray = shows.filter((show) => show.imdbID === nextState.params.id)
-    if (showArray.length < 1) {
-      return replace('/')
-    }
-    Object.assign(nextState.params, showArray[0])
-    return nextState
-  },
   render () {
     return (
       <Provider store={store}>
         <Router history={hashHistory}>
           <Route path='/' component={Layout}>
             <IndexRoute component={Landing} />
-            <Route path='/search' component={Search} shows={shows} />
-            <Route path='/details/:id' component={Details} onEnter={this.assignShow} />
+            <Route path='/search' component={Search} />
+            <Route path='/details/:id' component={Details} />
           </Route>
         </Router>
       </Provider>
@@ -33,4 +23,4 @@ const App = React.createClass({
   }
 })
 
-ReactDOM.render(<App />, document.getElementById('app'))
+module.exports = App
