@@ -2,7 +2,7 @@ const path = require('path')
 
 module.exports = {
   context: __dirname,
-  entry: './js/ClientApp.jsx',
+  entry: './js/BroswerEntry.jsx',
   output: {
     path: path.join(__dirname, '/public'),
     filename: 'bundle.js'
